+++ src/components/molecules/Breadcrumb/Breadcrumb.tsx
@@ -5,7 +5,7 @@ import {
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+} from "@/components/ui/DropdownMenu";
 import { cn } from "@/lib/utils";
 import Link from "next/link";
 import { useContext } from "react";
