+++ src/components/organisms/Header/Header.tsx
@@ -11,14 +11,14 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+} from "@/components/ui/DropdownMenu";
 import {
   Sheet,
   SheetContent,
   SheetDescription,
   SheetHeader,
   SheetTitle,
-} from "@/components/ui/sheet";
+} from "@/components/ui/Sheet";
 import { signout } from "@/features/auth/actions/signout";
 import { errorToast, successToast } from "@/lib/toast";
 import { accountContext } from "@/providers/account";
