+++ src/components/organisms/Menu/Menu.tsx
@@ -9,7 +9,7 @@ import {
   AccordionContent,
   AccordionItem,
   AccordionTrigger,
-} from "@/components/ui/accordion";
+} from "@/components/ui/Accordion";
 
 export const Menu = () => {
   const path = usePathname();
